доделанные курсы валют

+++ src/main/java/com/bsac/meowa/repo/BranchRateRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bsac.meowa.model.BranchRate;
 
-public interface BranchRateRepository extends JpaRepository<BranchRate, Long> {
+public interface BranchRateRepository extends JpaRepository<BranchRate, Integer> {
 }
